ViPER: convert FIREqualizer to float + recover its window offsets
FIREqualizer was the only remaining fixed-point code (Q25 int32 samples/gains, 64-bit MAC, >>25
multiply). Rewrote it as a mathematically-equivalent float/double port: gains stored linear (the 2^25
scale cancels), sum>>(j+1) -> sum/2^(j+1), Q25 multiply -> plain float multiply. Also recovered the
window-offset tables from the full constants dump (add +2^j {1..256}, sub -2^j {-1..-256}) — they were
zeroed placeholders the earlier pass couldn't extract. It's the unused FIR reference (the active EQ is
the already-float IIREqualizer); the engine now uses float/double throughout. Build green.

+++ app/src/main/cpp/viper/effects/FIREqualizer.h
@@ -5,61 +5,60 @@
 /**
  * FIREqualizer
  *
- * Faithful reconstruction of ViPER4Android v2505 `FIREqualizer`
+ * Faithful FLOAT reconstruction of ViPER4Android v2505 `FIREqualizer`
  * (mangled: _ZN12FIREqualizer...).
  *
- * This is a fixed-point (Q25), 10-band linear-phase equalizer that operates
- * directly on interleaved stereo Q.xx integer samples. It does NOT use FFT /
- * frequency sampling. Internally it maintains, for each of 9 dyadic scales, an
- * incrementally-updated running sum (box filter / moving average) over a
- * 256-deep delay line. The 10 band outputs are formed by telescoping the
- * differences between adjacent-scale moving averages, each weighted by the
- * band gain:
+ * A 10-band linear-phase equalizer that operates on interleaved stereo float
+ * samples (NOT FFT / frequency sampling). For each of 9 dyadic scales it keeps an
+ * incrementally-updated running sum (box filter / moving average) over a 256-deep
+ * delay line. The 10 band outputs telescope the differences between adjacent-scale
+ * moving averages, each weighted by its band gain:
  *
  *   y = g[0]*(x - avg[0])
  *     + sum_{j=1..8} g[j]*(avg[j-1] - avg[j])
  *     + g[9]*avg[8]
  *
- * where avg[j] = runningSum[j] >> (j+1) and x is the center-tap (delayed) input.
+ * where avg[j] = runningSum[j] / 2^(j+1) and x is the center-tap (delayed) input.
  *
- * All public method names/signatures match the original binary; nothing else in
- * the codebase currently calls this class.
+ * The original binary is Q25 fixed point; this is a mathematically-equivalent
+ * float port (the Q25 2^25 scale cancels out, so gains are stored linear and the
+ * multiplies/averages are plain float/double). Public method names/signatures
+ * match the original. Currently unused — the active EQ is viper::dsp::IIREqualizer.
  */
 class FIREqualizer {
 public:
     FIREqualizer();
     ~FIREqualizer();
 
-    // band: 0..9. level is a gain in "raw" units (see .cpp for the Q25 encoding).
+    // band: 0..9. level is a gain around unity (0 = flat); positive values are doubled.
     void SetBandLevel(uint32_t band, float level);
     void SetSamplingRate(uint32_t samplingRate);
     // Returns true if the enabled state actually changed.
     bool SetEnable(bool enable);
     void Reset();
 
-    // samples: interleaved stereo, fixed-point. frameCount = number of stereo frames.
-    void Process(int32_t *samples, int32_t frameCount);
+    // samples: interleaved stereo float. frameCount = number of stereo frames.
+    void Process(float *samples, int32_t frameCount);
 
 private:
     static constexpr int kHistorySize = 256;   // delay-line ring depth (indexed & 0xff)
     static constexpr int kStageCount = 9;       // dyadic moving-average scales -> 10 bands
 
-    // Per-band gains in Q25 (1.0 == 2^25). Stored reversed: band b -> mBandGains[9 - b].
-    int32_t mBandGains[10];
+    // Per-band linear gains around unity (1.0 == flat). Stored reversed: band b -> mBandGains[9 - b].
+    float mBandGains[10];
 
-    // Delay line holding the input samples already pushed out of the recent ring;
-    // the "subtract" stage reads from here to remove samples leaving each window.
-    int32_t mDelayedL[kHistorySize];
-    int32_t mDelayedR[kHistorySize];
+    // Delay line of samples already pushed out of the recent ring; the "subtract"
+    // stage reads it to remove samples leaving each window.
+    float mDelayedL[kHistorySize];
+    float mDelayedR[kHistorySize];
 
-    // Recent input ring; the "add" stage reads from here to add samples entering
-    // each window.
-    int32_t mInputL[kHistorySize];
-    int32_t mInputR[kHistorySize];
+    // Recent input ring; the "add" stage reads it to add samples entering each window.
+    float mInputL[kHistorySize];
+    float mInputR[kHistorySize];
 
-    // 64-bit running sums for each dyadic scale.
-    int64_t mAccumL[kStageCount];
-    int64_t mAccumR[kStageCount];
+    // Running sums per dyadic scale (double for accumulation precision).
+    double mAccumL[kStageCount];
+    double mAccumR[kStageCount];
 
     uint32_t mRingPos;
     uint32_t mSamplingRate;
